backport of commit 4e076a30309e85743a76c2b63d67e382d1712b85
